pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "copal-cli"
version = "0.1.0"
description = "CoPal - Command-line orchestration playbook for AI coding agents"
readme = "README.md"
requires-python = ">=3.9"
authors = [
    { name = "CoPal Team" }
]
keywords = ["ai", "cli", "developer-tools", "codex", "copilot", "claude"]
dependencies = [
    "PyYAML>=6.0",
    "jinja2>=3.1",
    "networkx>=3.0",
    "rich>=13.0",
    "InquirerPy>=0.3.4",
    "jsonschema>=4.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
]
vector-chroma = [
    "chromadb>=0.4",
]
vector-qdrant = [
    "qdrant-client>=1.7",
]

[project.scripts]
copal = "copal_cli.cli:main"

[tool.setuptools.packages.find]
where = ["."]
include = ["copal_cli", "copal_cli.*"]

[tool.setuptools.package-data]
copal_cli = [
    "templates/**/*",
    "skills/**/*",
]
